requirements.txt
common-code[test] @ git+https://github.com/swiss-ai-center/common-code.git@main
ultralytics==8.2.25
pillow==10.2.0
numpy==1.26.4
py-cpuinfo==9.0.0
